server/requirements.txt
requests==2.26.0
flask==2.0.2
gunicorn==20.1.0
python-dotenv==0.19.0
pytest==6.2.5
sentry-sdk[flask]==1.4.3
mediacloud==3.12.4
pandas==1.3.3
